@@ -5,6 +5,7 @@
 		"url": "https://github.com/FriendsOfTYPO3/tea.git"
 	},
 	"keywords": [],
+	"type": "module",
 	"author": "",
 	"version": "1.0.0",
 	"license": "ISC",
@@ -15,13 +16,13 @@
 		"lint:style:fix": "stylelint Resources/Public/**/*.css --fix"
 	},
 	"devDependencies": {
-		"eslint": "^8.57.0",
+		"eslint": "^9.2.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-import": "^2.29.1",
 		"eslint-plugin-prettier": "^5.1.3",
+		"install": "^0.13.0",
 		"prettier": "^3.2.5",
-		"stylelint": "^15.11.0",
-		"stylelint-config-recommended": "^13.0.0",
+		"stylelint": "^16.5.0",
+		"stylelint-config-standard-scss": "^13.1.0",
 		"stylelint-no-browser-hacks": "^1.3.0"
 	}
 }
